IT WORKS

+++ GameJam2023/Assets/Scriptable Objects/Scripts/UnitScriptableObject.cs
@@ -20,6 +20,6 @@ public class UnitScriptableObject : ScriptableObject
 
     [SerializeField] internal int attackPower;
 
-    [SerializeField] internal int range;
+    [SerializeField] internal float range;
 
 }
